requirements.txt
asn1crypto==0.24.0
certifi==2018.11.29
cffi==1.12.0
chardet==3.0.4
cryptography==2.5
future==0.17.1
idna==2.8
py-zabbix==1.1.5
pycparser==2.19
python-telegram-bot==11.1.0
requests==2.21.0
six==1.12.0
urllib3==1.24.1
